+++ core/src/main/java/org/svip/sbomfactory/serializers/SerializerFactory.java
@@ -139,10 +139,13 @@ public class SerializerFactory {
         Schema schema = CDX14;
         Format format = JSON;
 
-        if (fileContents.contains("SPDX")) schema = SPDX23;
+        if (fileContents.contains("SPDX") && fileContents.toLowerCase().contains("spdxversion")) schema = SPDX23;
         else if (fileContents.contains("rootComponent")) schema = SVIP; // Field unique to SVIP SBOM
 
-        if (fileContents.contains("DocumentName:")) format = TAGVALUE;
+        if (fileContents.contains("DocumentName:") || fileContents.contains("DocumentNamespace:")) {
+            schema = SPDX23;
+            format = TAGVALUE;
+        }
 
         // TODO what if we still have an incorrect deserializer?
         return schema.getDeserializer(format);